Update vtkTextMapper's alignment to match current conventions.
I missed this when I was updating the text actors to better support
rotations + alignments. Basically, now we anchor the texture's image
extent index <0, 0, 0> to the mapper/actor's position to get "proper"
alignment.

Also add a bit to ensure that this is exported properly by GL2PS and
prevent errors from converting NULL to std::string.

These features are tested by a test added in a later patch, in this
branch, as the other TextMapper implementation is still used by default
at this point.

Change-Id: Ia719e11a0598b601ac5a3437362d70a4bb00522f