Align to a fixed-height font bbox, rather than a "tight" one.
I used to move the alignment point onto the edge of the rendered
text, but this lead to some layout issues, including a particularly
nasty possibility of having a label rendered outside of a constrained
bounding box.

This should be made an option at some point (a tight alignment property
on vtkTextProperty perhaps?), as it does look nicer on certain centered
labels.

Change-Id: Idbbf97d1d7479c50d28ef3dd7db1f0e0928d32a4